backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.3.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
asyncpg==0.29.0          # PostgreSQL async driver (production)
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0

# IP Intelligence Module dependencies
tenacity==8.5.0  # Retry/backoff for external API calls
redis==5.0.1     # Caching layer (optional)
aiohttp==3.9.1   # Async HTTP client (alternative to httpx)
respx==0.21.1    # HTTPX mocking for tests
